+++ src/validators/local/acceptance.js
@@ -29,7 +29,3 @@ export const acceptanceLocalValidator = (element, options) => {
     return options.message
   }
 }
-
-export default {
-  acceptanceLocalValidator
-}